requirements.txt
django-environ==0.4.5
celery==4.4.7
Django==3.1
django-celery-beat==2.0.0
djangorestframework==3.11.1
flake8==3.8.3
psycopg2==2.8.5
requests==2.24.0
